refactor: excluindo arquivos individuais modCheckin para unificar em um único arquivo.
