--- src/modCheckin.h
@@ -1,6 +1,0 @@
-#ifndef MODULO_CHECKIN_H
-#define MODULO_CHECKIN_H
-
-char moduloCheckin(void);
-
-#endif
