--- src/ui/checkin/atualizarCheckin.c
@@ -1,18 +1,0 @@
-#include <stdio.h>
-
-char telaAtualizarCheckin(void) {
-
-    char op;
-    printf("====================\n");
-    printf("ATUALIZAR CHECKIN \n");
-    printf("====================\n");
-    printf("  ...EM BREVE... \n");
-    printf("\n");
-    printf("<Aperte 0 para sair!>\n");
-    
-    scanf("%c", &op);
-    getchar();
-
-    return op;
-
-    }
