--- src/ui/checkin/atualizarCheckin.h
@@ -1,6 +1,0 @@
-#ifndef ATUALIZAR_CHECKIN_H
-#define ATUALIZAR_CHECKIN_H
-
-char telaAtualizarCheckin(void);
-
-#endif
