--- src/ui/checkin/cadastrarCheckin.h
@@ -1,6 +1,0 @@
-#ifndef CADASTRAR_CHECKIN_H
-#define CADASTRAR_CHECKIN_H
-
-char telaCadastrarCheckin(void);
-
-#endif
