--- src/ui/checkin/excluirCheckin.c
@@ -1,19 +1,0 @@
-#include <stdio.h>
-
-char telaExcluirCheckin(void) {
-    
-    char op;    
-
-    
-    printf("====================\n");
-    printf("EXCLUIR CHECKIN\n");
-    printf("====================\n");
-    printf("  ...EM BREVE... \n");
-    printf("\n");
-    printf("<Aperte 0 para sair!>\n");
-
-    scanf("%c", &op);
-    getchar();
-
-    return op;
-}
