--- src/ui/checkin/excluirCheckin.h
@@ -1,6 +1,0 @@
-#ifndef EXCLUIR_CHECKIN_H
-#define EXCLUIR_CHECKIN_H
-
-char telaExcluirCheckin(void);
-
-#endif
