--- src/ui/checkin/telaCheckin.h
@@ -1,6 +1,0 @@
-#ifndef TELA_CHECKIN_H
-#define TELA_CHECKIN_H
-
-char telaCheckin(void);
-
-#endif
