--- src/ui/checkin/visualizarCheckin.h
@@ -1,6 +1,0 @@
-#ifndef VISUALIZAR_CHECKIN_H
-#define VISUALIZAR_CHECKIN_H
-
-char telaVisualizarCheckin(void);
-
-#endif
