+++ src/ui/telasCheckin.h
@@ -1,0 +1,11 @@
+#ifndef TELASCHECKIN_H
+#define TELASCHECKIN_H
+
+char moduloCheckin(void);
+char telaCheckin(void);
+char telaCadastrarCheckin(void);
+char telaVisualizarCheckin(void);
+char telaAtualizarCheckin(void);
+char telaExcluirCheckin(void);
+
+#endif
